Remove duplicate code after server CLI duplication - now every command in separate file
